pyproject.toml
[project]
name = "tdnet-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.78.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
